Script updated after first review iteration
